requirements.txt
streamlit
pandas
numpy
opencv-python-headless
tensorflow
pillow
